Hardened the code a bit by implementing checks from the solution file

+++ functions/get_files_info.py
@@ -56,10 +56,16 @@ def write_file(working_directory, file_path, content):
 
     if os.path.commonpath([abs_working_dir, abs_file_path]) != abs_working_dir:
         return f'Error: Cannot read "{file_path}" as it is outside the permitted working directory'
+    
+    if os.path.exists(abs_file_path) and os.path.isdir(abs_file_path):
+        return f'Error: "{file_path}" is a directory, not a file'
 
     head, _ = os.path.split(abs_file_path)
     if not os.path.exists(head):
-        os.makedirs(head)
+        try:
+            os.makedirs(head, exist_ok=True)
+        except Exception as e:
+            return f"Error: creating directory: {e}"
 
     try:
         with open(abs_file_path, "w") as f:
@@ -68,6 +74,28 @@ def write_file(working_directory, file_path, content):
         return f'Error writing file "{file_path}": {e}'
     
     return f'Successfully wrote to "{abs_file_path}" ({len(content)} characters written)'
+
+def write_file_alt(working_directory, file_path, content):
+
+    abs_working_dir = os.path.abspath(working_directory)
+    abs_file_path = os.path.abspath(os.path.join(working_directory, file_path))
+    if not abs_file_path.startswith(abs_working_dir):
+        return f'Error: Cannot write to "{file_path}" as it is outside the permitted working directory'
+    if not os.path.exists(abs_file_path):
+        try:
+            os.makedirs(os.path.dirname(abs_file_path), exist_ok=True)
+        except Exception as e:
+            return f"Error: creating directory: {e}"
+    if os.path.exists(abs_file_path) and os.path.isdir(abs_file_path):
+        return f'Error: "{file_path}" is a directory, not a file'
+    try:
+        with open(abs_file_path, "w") as f:
+            f.write(content)
+        return (
+            f'Successfully wrote to "{file_path}" ({len(content)} characters written)'
+        )
+    except Exception as e:
+        return f"Error: writing to file: {e}"
 
 def get_file_content_alt(working_directory, file_path):
